Taxify - Sprint 5

+++ src/taxify/IStatistics.java
@@ -6,10 +6,10 @@ public interface IStatistics {
     public int getReviews();
     public double getStars();
     public int getDistance();
-    public int getBilling();
+    public double getBilling();
     public void updateServices();
     public void updateReviews();
     public void updateStars(int stars);
     public void updateDistance(int distance);
-    public void updateBilling(int billing);
+    public void updateBilling(double billing);
 }
